TST #823 relocate
